+++ CHANGES.rst
@@ -4,6 +4,22 @@ Change Log
 Unreleased
 ----------
 
+**Added**
+
+* Add method :meth:`~.Reddit.patch` to :class:`.Reddit` class to support HTTP
+  PATCH requests.
+* Add class :class:`.Preferences` to access and update Reddit preferences.
+* Add attribute :attr:`.User.preferences` to access an instance of
+  :class:`.Preferences`.
+* Add method :meth:`.Message.delete()`.
+
+**Deprecated**
+
+* ``subreddit.submissions`` as the API endpoint backing the method is going
+  away. See
+  https://www.reddit.com/r/changelog/comments/7tus5f/update_to_search_api/.
+
+
 **Fixed**
 
 * Fix bug with positive ``pause_after`` values in streams provided by
@@ -11,7 +27,8 @@ Unreleased
   ``None``.
 * Parse URLs with trailing slashes and no ``'comments'`` element when creating
   :class:`.Submission` objects.
-* Fix bug where ``subreddit.submissions()`` returns a same submission more than once
+* Fix bug where ``subreddit.submissions()`` returns a same submission more than
+  once
 * Fix bug where ``ListingGenerator`` fetches the same batch of submissions in
   an infinite loop when ``'before'`` parameter is provided.